Added functionality for site administration by the manager.

+++ src/main/java/ua/koniukh/cargomanagementsystem/service/OrderService.java
@@ -1,32 +1,42 @@
 package ua.koniukh.cargomanagementsystem.service;
 
 import com.sun.istack.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import ua.koniukh.cargomanagementsystem.model.Cargo;
 import ua.koniukh.cargomanagementsystem.model.Order;
+import ua.koniukh.cargomanagementsystem.model.User;
 import ua.koniukh.cargomanagementsystem.model.dto.OrderDTO;
+import ua.koniukh.cargomanagementsystem.repository.CargoRepository;
 import ua.koniukh.cargomanagementsystem.repository.OrderRepository;
+import ua.koniukh.cargomanagementsystem.repository.UserRepository;
 
-import java.math.BigDecimal;
+import javax.transaction.Transactional;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class OrderService {
 
+    private final CargoRepository cargoRepository;
+    private final UserRepository userRepository;
     private final OrderRepository orderRepository;
     private final UserService userService;
     private final PriceService priceService;
     private final CargoService cargoService;
 
-    public OrderService(OrderRepository orderRepository, UserService userService, PriceService priceService, CargoService cargoService) {
+    public OrderService(CargoRepository cargoRepository, UserRepository userRepository, OrderRepository orderRepository, UserService userService, PriceService priceService, CargoService cargoService) {
+        this.cargoRepository = cargoRepository;
+        this.userRepository = userRepository;
         this.orderRepository = orderRepository;
         this.userService = userService;
         this.priceService = priceService;
         this.cargoService = cargoService;
+    }
+
+    public List<Order> showApprove() {
+        return orderRepository.findAllByApprovedIsTrue();
     }
 
     public Order findById(Long id) {
@@ -66,6 +76,7 @@ public class OrderService {
                 .cargo(null)
                 .date(LocalDate.parse(orderDTO.getDate()))
                 .packing(orderDTO.isPacking())
+                .approved(false)
                 .declaredValue(orderDTO.getDeclaredValue())
                 .price(priceService.calculatePrice(orderDTO))
                 .type(orderDTO.getType())
@@ -79,10 +90,37 @@ public class OrderService {
         return order;
     }
 
+    //todo write own exeption
+    @Transactional
     public void deleteById(Long id) {
-        if (findById(id).getCargo().getId() != 0) {
-            cargoService.deleteById(findById(id).getCargo().getId());
+        try {
+           if (cargoRepository.existsById(findById(id).getCargo().getId())) {
+               cargoService.deleteById(findById(id).getCargo().getId());
+           }
+        } catch (NullPointerException e) {
+            e.printStackTrace();
         }
+        if (orderRepository.existsById(id))
         orderRepository.deleteById(id);
+    }
+
+    @Transactional
+    public void deleteUserById(Long id) {
+        User user = userService.findById(id);
+        List<Order> list = new ArrayList<>(user.getOrders());
+        if (list.isEmpty()) {
+            userRepository.deleteById(id);
+        }
+        for (Order order : list) {
+            deleteById(order.getId());
+        }
+        userRepository.deleteById(id);
+    }
+
+    @Transactional
+    public void approveOrder(Long id) {
+        Order order = findById(id);
+        order.setApproved(true);
+        saveOrder(order);
     }
 }
